ambient-light/AmbientLightSensor-supported-by-feature-policy.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('ambient-light-sensor', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise ambient-light-sensor.');
</script>
